删除新用户,生成邀请码,生成新邀请记录,行事日查询,日历的动态修改数据,生成访问展区记录功能实现,邀请码,访问记录相关数据库表字段建立

+++ src/main/java/com/example/admin/mapper/UserMapper.java
@@ -1,14 +1,15 @@
 package com.example.admin.mapper;
 import com.example.admin.entity.User;
-import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
+
+import java.util.List;
 
 @Mapper
 public interface UserMapper {
     List<User> findAll();
     List<User> findbyNamePage(int limit, int offset, String name);
-    User findById(Long id);
+    User findById(int id);
     void insert(User user);
     void update(User user);
-    void delete(Long id);
+    void delete(int id);
 }
